+++ Modules/Changelog/Changelog.lua
@@ -1,6 +1,72 @@
 local  _, CLM = ...
-
+-- ### Auto trade
+-- Auto trade for soulbound items with trade timer not exipred should now work properly.
 CLM.ChangelogData = {
+    {
+        version = "v1.4.0",
+        data = {
+            {
+                name = "Developers note",
+                data = {
+                    {header = "CLM Discord", body = "Join our discord for more info: |cff44cc44https://tiny.one/clm-discord|r"},
+                    {header = "Mount Hyjal and Black Temple Raids", body = "Good luck everyone with your progress, and may the loot ever be in your favor."},
+                },
+            },
+            {
+                name = "Translations",
+                data = {
+                    {header = "French", body = "First version of French translation available thanks to |cff44cc44Paracétamole|r."},
+                    {header = "Russian", body = "Minor updates as always thanks to cff44cc44PCyber|r."},
+                },
+            },
+            {
+                name = "Features",
+                data = {
+                    {header = "Raid Messages", body = "You can now disable specific messages so they won't be displayed in Raid Warnings or Raid chat. Thank you |cff44cc44grecaun|r."},
+                },
+            },
+            {
+                name = "Fixes",
+                data = {
+                    {header = "Auctioning", body = "You are no longer able to bid base value if you already bid in Open Ascending auction."},
+                    {header = "Loot announcement", body = "Fixed announcement of dropped loot from corpse. Thank you |cff44cc44grecaun|r."},
+                    {header = "Auto trade", body = "Auto trade for soulbound items with trade timer not exipred should now work properly."},
+                },
+            },
+        }
+    },
+    {
+        version = "v1.3.1",
+        data = {
+            {
+                name = "Developers note",
+                data = {
+                    {header = "CLM Discord", body = "Join our discord for more info: |cff44cc44https://tiny.one/clm-discord|r"},
+                    {header = "Happy new Year!", body = "It has been a year. Our journey with CLM began in January 2021. Thank you all and don't hesitate to |cff44cc44support us|r!"},
+                },
+            },
+            {
+                name = "Features",
+                data = {
+                    {header = "Attendance tracking", body = "CLM now tracks and reports attendance based on local configuration. You can set 2 parameters: |cffcccc00Max raids per week|r and |cffcccc00Week window size|r. First one defines the amount of raiders a player must join to have |cffcccc00100%|r attendance. The latter determines the number of weeks take into account."},
+                    {header = "Auto Item Award and Auto Trade", body = "CLM will now auto-give items if awarding them when corpse is open, or will try to auto-trade them otherwise. Both behaviors can be disabled in configuration."},
+                    {header = "Extended auction information", body = "Addon tracks now various additional bidder behavior: If they did pass, can't use item, closed window without any action or let the timer expire. You can now get those various informations by hovering over |cffcccc00Info|r icon. |cff44cc44PASS|r has been removed from main bid display. No action might also mean player has not received bid command (was during reload) or has no or outdated addon installed."},
+                    {header = "Open auction alignment", body = "Players can no longer |cffcc0000cancel|r bids in Open Auctions or |cffcccc00pass|r if they already bid. Players can now always bid minimum in those auctions. This allows for handling both English and Swedish House Auctions."},
+                    {header = "Item notes", body = "Item notes are now stored locally per item (they are never synced)."},
+                    {header = "New filtering options", body = "New filters are available in various UIs: Mains only; In Guild only; External (outside of guild) only"},
+                },
+            },
+            {
+                name = "Fixes",
+                data = {
+                    {header = "Bid sorting", body = "Bids will now properly sort in the order of Highest Bid 1st > Total DKP 2nd (on equal bids)."},
+                    {header = "Class sorting", body = "Classess in various UIs will now properly sort alphabetically."},
+                    {header = "Icon fixes", body = "Missing CLM Icons should now be visible again."},
+                    {header = "Localisation enhancements", body = "Multiple localisation enhancements added."},
+                },
+            },
+        }
+    },
     {
         version = "v1.3.0",
         data = {